Add files via upload
This program lets you choose between encoding or decoding by taking an input message and a shift amount and then taking each letter in the message and adding the shift amount to it. a shifted 5 becomes e. Decoding does the same with a reversed shift to produce the original message.

ex:
encode: hal shifted 1 = ibm
decode: ibm shifted 1 = hal